Abstracted instruments to store
