+++ src/stores/index.ts
@@ -1,4 +1,5 @@
 export { default as highlightedNotes } from './highlightedNotes';
+export { default as instruments } from './instruments';
 export { default as mode } from './mode';
 export { default as root } from './root';
 export { default as selectedNote } from './selectedNote';
